Abstracting player and match.
