+++ src/main/java/entities/player/MatchHostEntity.java
@@ -1,0 +1,14 @@
+package entities.player;
+
+import model.player.MatchPlayerEntity;
+import model.connection.Connection;
+
+public final class MatchHostEntity extends MatchPlayerEntity
+{
+
+    public MatchHostEntity(Connection i_Connection)
+    {
+        super(i_Connection);
+    }
+
+}
