--- src/main/java/match/entities/match_host/MatchHostEntity.java
@@ -1,14 +1,0 @@
-package match.entities.match_host;
-
-import match.entities.MatchEntity;
-import servers.connection.Connection;
-
-public final class MatchHostEntity extends MatchEntity
-{
-
-    public MatchHostEntity(Connection i_Connection)
-    {
-        super(i_Connection);
-    }
-
-}
