--- src/main/java/servers/connection/webconnection/WebConnection.java
@@ -1,13 +1,0 @@
-package servers.connection.webconnection;
-
-
-public final class WebConnection {
-
-
-}
-
-
-
-
-
-
